requirements.txt
appdirs==1.4.4
asn1crypto==1.3.0
click==7.1.2
Flask==1.1.2
Flask-JWT-Extended==3.24.1
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
nose==1.3.7
numpy==1.18.5
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==2.8.1
six==1.15.0
Werkzeug==1.0.1
wrapt==1.12.1
